+++ src/CAAS/Models/Symmetric/Encryption/SymmetricEncryptionResponse.cs
@@ -13,6 +13,6 @@ namespace CAAS.Models.Symmetric.Encryption
 
         [Required]
         [DefaultValue(11)]
-        public long ProcessingTimeInMs { get; set; }
+        public double ProcessingTimeInMs { get; set; }
     }
 }
